@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
-    <LangVersion>9</LangVersion>
+    <LangVersion>12</LangVersion>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Authors>GreemDev</Authors>
     <Company>GreemDev</Company>
@@ -11,9 +11,9 @@
     <RepositoryUrl>https://github.com/GreemDev/Gommon</RepositoryUrl>
     <Description>Common classes/utilities used in my projects. Made for redundancy so I don't have duplicate code in my repositories.</Description>
     <NeutralLanguage>en-US</NeutralLanguage>
-    <Version>2.5.2</Version>
+    <Version>2.6.0</Version>
     <PackageReleaseNotes>
-      Simplify code
+      Update C# version, more simplifications and improvements.
     </PackageReleaseNotes>
   </PropertyGroup>
 
