@@ -4,59 +4,59 @@ using System.Runtime.CompilerServices;
 using JetBrains.Annotations;
 
 // ReSharper disable MemberCanBePrivate.Global
-namespace Gommon {
-    public static class Collections {
-        //this might look dumb, but it's basically to avoid new[]{ values } or new {CollectionType}<T>() { values }
-        //you could also `using static` this class and call its methods without the class name, NewList(values)
+namespace Gommon;
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static List<T> NewList<T>([ItemCanBeNull] params T[] values) => new(values);
+public static class Collections {
+    //this might look dumb, but it's basically to avoid new[]{ values } or new {CollectionType}<T>() { values }
+    //you could also `using static` this class and call its methods without the class name, NewList(values)
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static HashSet<T> NewHashSet<T>([ItemCanBeNull] params T[] values) => new(values);
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static List<T> NewList<T>([ItemCanBeNull] params T[] values) => [..values];
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Queue<T> NewQueue<T>([ItemCanBeNull] params T[] initialValues) => new(initialValues);
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static HashSet<T> NewHashSet<T>([ItemCanBeNull] params T[] values) => [..values];
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static LinkedList<T> NewLinkedList<T>([ItemCanBeNull] params T[] values) => new(values);
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static Queue<T> NewQueue<T>([ItemCanBeNull] params T[] initialValues) => new(initialValues);
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Stack<T> NewStack<T>([ItemCanBeNull] params T[] values) => new(values);
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static LinkedList<T> NewLinkedList<T>([ItemCanBeNull] params T[] values) => new(values);
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static T[] NewArray<T>([ItemCanBeNull] params T[] values) => values;
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static Stack<T> NewStack<T>([ItemCanBeNull] params T[] values) => new(values);
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static Dictionary<TKey, TValue> NewDictionary<TKey, TValue>(params (TKey Key, TValue Value)[] pairs)
-            => pairs.ToDictionary(x => x.Key, x => x.Value);
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static T[] NewArray<T>([ItemCanBeNull] params T[] values) => values;
 
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static SafeDictionary<TKey, TValue> NewSafeDictionary<TKey, TValue>(
-            params (TKey Key, TValue Value)[] pairs) where TValue : class
-            => NewDictionary(pairs).AsSafe();
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static Dictionary<TKey, TValue> NewDictionary<TKey, TValue>(params (TKey Key, TValue Value)[] pairs)
+        => pairs.ToDictionary(x => x.Key, x => x.Value);
 
-        public static SafeDictionary<TKey, TValue> AsSafe<TKey, TValue>(
-            [CanBeNull] this Dictionary<TKey, TValue> current)
-            where TValue : class
-            => new(current ?? new Dictionary<TKey, TValue>());
-    }
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static SafeDictionary<TKey, TValue> NewSafeDictionary<TKey, TValue>(
+        params (TKey Key, TValue Value)[] pairs) where TValue : class
+        => NewDictionary(pairs).AsSafe();
 
-    /// <summary>
-    ///     Mutable alternative <see cref="Dictionary{TKey,TValue}"/> implementation solely for nullable indexing.
-    /// </summary>
-    /// <typeparam name="TKey">Entry key</typeparam>
-    /// <typeparam name="TValue">Entry value</typeparam>
-    public class SafeDictionary<TKey, TValue> : Dictionary<TKey, TValue> where TValue : class {
-        public SafeDictionary(Dictionary<TKey, TValue> initial) : base(initial.Count)
-            => initial.ForEach(x => Add(x.Key, x.Value));
+    public static SafeDictionary<TKey, TValue> AsSafe<TKey, TValue>(
+        [CanBeNull] this Dictionary<TKey, TValue> current)
+        where TValue : class
+        => new(current ?? new Dictionary<TKey, TValue>());
+}
+
+/// <summary>
+///     Mutable alternative <see cref="Dictionary{TKey,TValue}"/> implementation solely for nullable indexing.
+/// </summary>
+/// <typeparam name="TKey">Entry key</typeparam>
+/// <typeparam name="TValue">Entry value</typeparam>
+public class SafeDictionary<TKey, TValue> : Dictionary<TKey, TValue> where TValue : class {
+    public SafeDictionary(Dictionary<TKey, TValue> initial) : base(initial.Count)
+        => initial.ForEach(x => Add(x.Key, x.Value));
         
-        public SafeDictionary() {}
+    public SafeDictionary() {}
 
-        [CanBeNull]
-        public new TValue this[TKey key] {
-            get => this.FindValue(key).OrDefault();
-            set => base[key] = value;
-        }
+    [CanBeNull]
+    public new TValue this[TKey key] {
+        get => this.FindValue(key).OrDefault();
+        set => base[key] = value;
     }
 }
